Move half destroy logic to the dynamic slab hooks
